requirements.txt
# local package
-e .

# external requirements
#Sphinx
#coverage
#awscli
#flake8
click==8.1.3
python-dotenv>=0.5.1
tensorflow~=2.9
keras~=2.10.0
scikit-learn~=1.1.2
numpy~=1.23.4
opencv-python~=4.6.0.66
numpy~=1.23.4
matplotlib
pandas==1.5.0
split-folders==0.5.1
mlflow==1.29.0
dvc
tensorflow_hub
split-folders
random2
glob2
pytest
pylint
great_expectations
fastapi[all]